+++ chat-widget/src/components/webchatcontainerstateful/hooks/usePersistentChatHistory.ts
@@ -1,0 +1,32 @@
+import { useEffect, useRef } from "react";
+
+import ChatWidgetEvents from "../../livechatwidget/common/ChatWidgetEvents";
+import { FacadeChatSDK } from "../../../common/facades/FacadeChatSDK";
+import { IPersistentChatHistoryProps } from "../../livechatwidget/interfaces/IPersistentChatHistoryProps";
+import PersistentConversationHandler from "../../livechatwidget/common/PersistentConversationHandler";
+
+const usePersistentChatHistory = (facadeChatSDK: FacadeChatSDK | undefined, props : IPersistentChatHistoryProps) => {
+    const handlerRef = useRef<PersistentConversationHandler | null>(null);
+
+    useEffect(() => {
+        if (!facadeChatSDK) {
+            return;
+        }
+
+        // Initialize a new instance of PersistentConversationHandler
+        handlerRef.current = new PersistentConversationHandler(facadeChatSDK, props);
+
+        const handler = async () => {
+            await handlerRef.current?.pullHistory();
+        };
+
+        window.addEventListener(ChatWidgetEvents.FETCH_PERSISTENT_CHAT_HISTORY, handler);
+
+        return () => {
+            window.removeEventListener(ChatWidgetEvents.FETCH_PERSISTENT_CHAT_HISTORY, handler);
+            handlerRef.current?.reset(); // Reset the handler state on cleanup
+        };
+    }, [facadeChatSDK]);
+};
+
+export default usePersistentChatHistory;