+++ chat-widget/src/components/webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/activities/Constants.ts
@@ -1,0 +1,4 @@
+
+export const LazyLoadActivityConstants = {
+    SCROLL_ID: ".webchat__basic-transcript__scrollable"
+};