+++ chat-widget/src/components/webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/activities/ConversationDividerActivity.tsx
@@ -1,0 +1,14 @@
+import { IPersistentChatHistoryProps } from "../../../../../livechatwidget/interfaces/IPersistentChatHistoryProps";
+import React from "react";
+import { defaultPersistentChatHistoryProps } from "../../../../../livechatwidget/common/defaultProps/defaultPersistentChatHistoryProps";
+import { mergeStyles } from "@fluentui/react";
+
+const ConversationDividerActivity = (props: IPersistentChatHistoryProps) => {
+    const styleApplied = mergeStyles(
+        defaultPersistentChatHistoryProps.dividerActivityStyle,
+        props.dividerActivityStyle
+    );
+    
+    return <div className={styleApplied} />;
+};
+export default ConversationDividerActivity;